Extract logic to dynamically render a native dialog
